Refactor: profile 이미지 불러오는 부분 훅 분리
