+++ src/components/common/Profile.js
@@ -1,81 +1,53 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React, { useState } from 'react';
 import NotificationList from '../notifications/NotificationList';
-import decodeJwtResponse from '../../utils/decodeJwtResponse';
+import { useProfileImage } from '../../hooks/useProfileImage';
+
+const mockNotifications = [
+  {
+    id: 1,
+    content: 'ㅇㅇㅇ님이 습관 성공',
+    date: '2023-09-12',
+    time: '14:30',
+    status: 'success',
+    isVisible: true,
+  },
+  {
+    id: 2,
+    content: 'ㅇㅇㅇ님이 습관 실패',
+    date: '2023-09-12',
+    time: '14:30',
+    status: 'failure',
+    isVisible: true,
+  },
+  {
+    id: 3,
+    content: 'ㅇㅇㅇ님이 인증 요청',
+    date: '2023-09-12',
+    time: '14:30',
+    status: 'verificationRequest',
+    isVisible: true,
+  },
+  {
+    id: 4,
+    content: 'ㅇㅇㅇ님이 승인 완료',
+    date: '2023-09-12',
+    time: '14:30',
+    status: 'approveRequest',
+    isVisible: true,
+  },
+  {
+    id: 5,
+    content: 'ㅇㅇㅇ님이 초대',
+    date: '2023-09-12',
+    time: '14:30',
+    status: 'invite',
+    isVisible: true,
+  },
+];
 
 const Profile = () => {
   const [showNotifications, setShowNotifications] = useState(false);
-  const [profileImageUrl, setProfileImageUrl] = useState(null);
-
-  const accessToken = localStorage.getItem('accessToken');
-
-  useEffect(() => {
-    const fetchData = async () => {
-      let userId = null;
-      try {
-        const decoded = decodeJwtResponse(accessToken);
-        userId = decoded.userId;
-      } catch (error) {
-        console.error('Error decoding JWT:', error);
-        return;
-      }
-
-      try {
-        const response = await axios.get(
-          `${process.env.REACT_APP_SERVER_DOMAIN}/api/user/${userId}`,
-        );
-        const user = response.data;
-        setProfileImageUrl(user.profileImageUrl);
-      } catch (error) {
-        console.error('Error fetching user:', error);
-      }
-    };
-
-    fetchData();
-  }, [accessToken]);
-
-  const mockNotifications = [
-    {
-      id: 1,
-      content: 'ㅇㅇㅇ님이 습관 성공',
-      date: '2023-09-12',
-      time: '14:30',
-      status: 'success',
-      isVisible: true,
-    },
-    {
-      id: 2,
-      content: 'ㅇㅇㅇ님이 습관 실패',
-      date: '2023-09-12',
-      time: '14:30',
-      status: 'failure',
-      isVisible: true,
-    },
-    {
-      id: 3,
-      content: 'ㅇㅇㅇ님이 인증 요청',
-      date: '2023-09-12',
-      time: '14:30',
-      status: 'verificationRequest',
-      isVisible: true,
-    },
-    {
-      id: 4,
-      content: 'ㅇㅇㅇ님이 승인 완료',
-      date: '2023-09-12',
-      time: '14:30',
-      status: 'approveRequest',
-      isVisible: true,
-    },
-    {
-      id: 5,
-      content: 'ㅇㅇㅇ님이 초대',
-      date: '2023-09-12',
-      time: '14:30',
-      status: 'invite',
-      isVisible: true,
-    },
-  ];
+  const { profileImageUrl, error } = useProfileImage();
 
   const [notifications, setNotifications] = useState(mockNotifications);
 
@@ -84,7 +56,11 @@ const Profile = () => {
   return (
     <div className='relative' style={{ fontFamily: 'NotoSansKR' }}>
       <div onClick={() => setShowNotifications(!showNotifications)}>
-        {profileImageUrl ? (
+        {error ? (
+          <div className='bg-red-500 rounded w-[40px] h-[40px] text-white flex justify-center items-center'>
+            !
+          </div>
+        ) : profileImageUrl ? (
           <img
             src={profileImageUrl}
             alt='프로필 이미지'
